- Added commit_ref instead of commit_hash - Saving to gym/envs/custom/username/commit_hash - Restricted upload of custom envs to scoreboard - Added various warnings -

+++ gym/versioning/manager.py
@@ -1,7 +1,6 @@
 import json
 import logging
 import requests
-import shutil
 import os
 import gym
 import six.moves.urllib as urllib
@@ -9,14 +8,14 @@ from gym.envs import register, deregister
 
 from gym.versioning.config_parser import parse_config
 from gym.scoreboard.client import http_client, util
-from gym.scoreboard.client.resource import UserEnvConfig
+from gym.scoreboard import UserEnvConfig
 
 logger = logging.getLogger(__name__)
 
 verify_ssl_certs = True # [SECURITY CRITICAL] only turn this off while debugging
 http_client = http_client.RequestsClient(verify_ssl_certs=verify_ssl_certs)
 gym_abs_path = os.path.dirname(os.path.abspath(gym.__file__))
-user_envs_cache_path = 'envs/.user_envs.json'
+user_env_cache_name = '.envs.json'
 
 class VersioningManager(object):
     """
@@ -25,15 +24,18 @@ class VersioningManager(object):
     def __init__(self):
         self.user_envs = []
         self.env_ids = set()
-        self.cache_path = os.path.join(gym_abs_path, user_envs_cache_path)
+        self.custom_env_path = os.path.join(gym_abs_path, 'envs', 'custom')
 
     def load_user_envs(self):
         """ Loads downloaded user envs from filesystem cache on `import gym` """
-        if not os.path.exists(self.cache_path):
+        if not os.path.isdir(self.custom_env_path):
             return
-        with open(self.cache_path) as cache:
-            for line in cache:
-                self._load_env(line.rstrip('\n'))
+        for directory in os.listdir(self.custom_env_path):
+            current_cache_path = os.path.join(self.custom_env_path, directory, user_env_cache_name)
+            if os.path.isfile(current_cache_path):
+                with open(current_cache_path) as cache:
+                    for line in cache:
+                        self._load_env(current_cache_path, line.rstrip('\n'))
         if len(self.env_ids) > 0:
             logger.info('Found and registered %d user environments.', len(self.env_ids))
 
@@ -84,9 +86,9 @@ The repository must have a ".openai.yml" in its top-level folder on its master b
             else:
                 raise err
 
-        # Resetting username and repository if they were not included in config file
-        config['username'] = username if 'username' not in config else config['username']
-        config['repository'] = repository if 'repository' not in config else config['repository']
+        # Setting username and repository on config top-level
+        config['username'] = username
+        config['repository'] = repository
 
         # Parsing (might return more than one env if target_env == '*')
         parsed_envs = parse_config(config, target_env, target_version)
@@ -95,20 +97,22 @@ The repository must have a ".openai.yml" in its top-level folder on its master b
 
         # Downloading, and registering
         for parsed_env in parsed_envs:
-            env_prefix = parsed_env['id'].replace('-', '_').lower() # username/envname_v0
-            env_root_path = os.path.join(gym_abs_path, 'envs', env_prefix)
-
-            if os.path.exists(env_root_path):
-                shutil.rmtree(env_root_path)
+            env_prefix = '{}/{}'.format(parsed_env['username'], parsed_env['commit_hash']).replace('-', '_').lower() # username/commit_hash
+            env_root_path = os.path.join(self.custom_env_path, env_prefix)
 
             for env_file in parsed_env['files']:
                 self._download_user_env_file(parsed_env, env_file, env_root_path)
 
             # Adding missing __init__ files
-            if not os.path.isfile(os.path.join(gym_abs_path, 'envs', parsed_env['username'], '__init__.py')):
-                open(os.path.join(gym_abs_path, 'envs', parsed_env['username'], '__init__.py'), 'w').close()
-            if not os.path.isfile(os.path.join(env_root_path, '__init__.py')):
-                open(os.path.join(env_root_path, '__init__.py'), 'w').close()
+            target_files = [
+                os.path.join(self.custom_env_path, '__init__.py'),
+                os.path.join(self.custom_env_path, parsed_env['username'], '__init__.py'),
+                os.path.join(self.custom_env_path, parsed_env['username'], parsed_env['commit_hash'], '__init__.py')
+            ]
+            for target_file in target_files:
+                current_target = target_file.replace('-', '_').lower()
+                if not os.path.isfile(current_target):
+                    open(current_target, 'w').close()
 
             if self._register(parsed_env):
                 logger.info('Registered the environment: "%s"', parsed_env['id'])
@@ -125,6 +129,8 @@ The repository must have a ".openai.yml" in its top-level folder on its master b
             urllib.parse.quote_plus(util.utf8(env_file)).replace('%2F', '/'))
         target_file = os.path.join(env_root_path, env_file)
 
+        if os.path.isfile(target_file):
+            return  # Already cached for this commit hash, no need to re-download
         if not os.path.exists(os.path.dirname(target_file)):
             os.makedirs(os.path.dirname(target_file))
         with open(target_file, 'wb') as handle:
@@ -139,6 +145,9 @@ The repository must have a ".openai.yml" in its top-level folder on its master b
                 logger.warn('Unable to download file "%s" for env "%s". The env might not work properly and '
                             'you might need to re-pull it.\nRequest URL: %s\nResponse Code:%d',
                             env_file, user_env['id'], target_file, response.status_code)
+                # Removing possibly corrupt file from cache
+                if os.path.isfile(target_file):
+                    os.remove(target_file)
                 return
 
             for block in response.iter_content(1024):
@@ -163,15 +172,22 @@ The repository must have a ".openai.yml" in its top-level folder on its master b
         return True
 
     def _update_cache(self):
-        if os.path.exists(self.cache_path + '.bak'):
-            os.remove(self.cache_path + '.bak')
-        if os.path.exists(self.cache_path):
-            os.rename(self.cache_path, self.cache_path + '.bak')
-        with open(self.cache_path, 'w') as cache:
-            for user_env in self.user_envs:
-                cache.write('{}\n'.format(json.dumps(user_env)))
+        envs_by_username = {}
+        for user_env in self.user_envs:
+            username = user_env['username'].replace('-', '_').lower()
+            if username not in envs_by_username:
+                envs_by_username[username] = []
+            envs_by_username[username].append(user_env)
 
-    def _load_env(self, json_line):
+        for username in envs_by_username:
+            cache_file = os.path.join(self.custom_env_path, username, user_env_cache_name)
+            if not os.path.exists(os.path.dirname(cache_file)):
+                os.makedirs(os.path.dirname(cache_file))
+            with open(cache_file, 'w') as cache:
+                for user_env in envs_by_username[username]:
+                    cache.write('{}\n'.format(json.dumps(user_env)))
+
+    def _load_env(self, current_cache_path, json_line):
         if len(json_line) == 0:
             return
 
@@ -184,7 +200,7 @@ The repository must have a ".openai.yml" in its top-level folder on its master b
 
         if not valid_json or not 'id' in user_env:
             logger.warn('Unable to load user environment. Try deleting your cache '
-                        'file "%s" if this problem persists. \n\nLine: %s', self.cache_path, json_line)
+                        'file "%s" if this problem persists. \n\nLine: %s', current_cache_path, json_line)
             return None
 
         self._register(user_env)
